Implement Continuous Integration, running tests with Github Actions
The list of packages installed splits between those necessary for
testing and those necessary for getting the tests to run to completion.

For the purposes of this patch series, "The tests" are defined as only
what is run by `nosetests`, as already configured in this repository.

This patch does NOT install the Ontospy package to pick up dependencies,
nor does it use `requirements.txt`. A follow-on patch will do a package
installation for dependency review, but that will necessitate another
patch related to JSON-LD be merged.  This first patch is intended to
demonstrate that unit tests can function through a CI system.

This patch series only runs CI for the latest Python version Ontospy
supports (per setup.py), Python 3.7.  This is due to a few reasons:
* setup.py reports support up through Python 3.7.
* The active list of Python versions includes back to Python 3.6,
  albeit, that is only through December of this year:
  https://devguide.python.org/#status-of-python-branches
* rdflib recently dropped support for Python 3.6, and a patch coming in
  this patch series would fail if tested on 3.6.

So, the intersection of supported and supportable Python versions ends
up being only Python 3.7.  Support (or at least, testing) of other
Python versions is left for separate discussion in another patch series.

Disclaimer:
Participation by NIST in the creation of the documentation of mentioned
software is not intended to imply a recommendation or endorsement by the
National Institute of Standards and Technology, nor is it intended to
imply that any specific software is necessarily the best available for
the purpose.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>