Have Continuous Integration install package locally
This patch sets the CI directions to pick up all dependencies from
installing the Ontospy package, rather than using a list separate from
`requirements.txt`.

This patch is known to trigger a CI failure.  A follow-on patch will
import the fix from another patch series.

Disclaimer:
Participation by NIST in the creation of the documentation of mentioned
software is not intended to imply a recommendation or endorsement by the
National Institute of Standards and Technology, nor is it intended to
imply that any specific software is necessarily the best available for
the purpose.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>