Restore .instances populate-on-first-get behavior
This was commented out in the patch just prior.  It was not immediately
clear from the patch documentation why this happened.  From the patch
just prior and onward, `nosetests` reports a failure from something
attempts to treat `.instances` like a list and encounters a `bool`
instead.

This patch reverts the comment-out, solely to get unit tests to pass
once again.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>