Adjust notion of "Top-level property" to account for SHACL shapes
This patch addresses a use case where an ontology
defines no or few rdfs:domain statements for its
properties, instead relating properties to classes
via SHACL shapes.  If a property is associated to
classes via SHACL property shapes within an ontology,
it does not seem appropriate to treat the property as
universally applicable.

Co-authored-by: Alex Nelson <alexander.nelson@nist.gov>
Signed-off-by: TJ Balon <tbalon@mitre.org>