Have class-shape confirmation function infer targetClass
The SHACL specification, section 2.1.3.3 "Implicit Class Targets",
indicates a reflexive sh:targetClass triple is implied for a SHACL Class
that is also a Shape.  The function updated in this patch reviewed
NodeShapes for being classes based on an explicit sh:targetClass
statement.  This patch changes the check to rely on that triple being
implied under the SHACL specification conditions.

Additionally, the "Class" definition is adjusted to being a SHACL Class,
which covers RDFS Classes in addition to the prior implementation of
only OWL classes.

References:
* https://www.w3.org/TR/shacl/#implicit-targetClass

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>