Expand tested Python versions
Python 3.7 is now past end of support.  The current versions available,
and their end of support dates, are listed on the python.org downloads
page.

This patch meant to expand testing to run CI against the oldest and most
recent active Python releases.  However, `nose` (not updated since 2015)
fails to run due to referencing `collections.Callable`, a construct
present in Python 2.7, but absent from documentation since Python 3.3,
and unable to be referenced starting in Python 3.10.

References:
* https://docs.python.org/2.7/library/collections.html
* https://docs.python.org/3.2/library/collections.html
* https://docs.python.org/3.3/library/collections.html
* https://www.python.org/downloads/

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>