Don't use linebreaks filter in Markdown templates (#64)
The Markdown Django templates were using the linebreaks filter on some
of the content.  This filter wraps content using <p> and <br />.
However this means that if the ontology was using Markdown in their
comments this would not get rendered as most Markdown renderers don't
apply rendering inside of HTML block elements like <p>.  It was also
unnecessary since Markdown renderers already use line breaks to suitably
render paragraphs.

This was most applicable for descriptions of things since those are the
places were ontology authors were most likely to want to use Markdown to
explain the intended usage of their ontology concepts.

By removing the use of this filter the ontology descriptions are output
as-is in the generated Markdown files allowing any Markdown in them to
be rendered by whatever Markdown renderer the user ultimately chooses.