Fix non-working tables in markdown
In current markdown template, in the "Instances of" section, the current
implementation is rendered as "code block" by github and by pandoc.

I think this is because of the indentation of the html bits. html tags
and special characters are escaped by pandoc, i.e. the html source reads

```
table border="1" cellspacing="3" cellpadding="5" class="classproperties table-hover ">
<pre><code>&lt;tr&gt;
    &lt;th height=&quot;40&quot;&gt;Property&lt;/th&gt;&lt;th&gt;Description&lt;/th&gt;&lt;th&gt;Expected Type&lt;/th&gt;
&lt;/tr&gt;

[etc...]
```

I have reindented the template file so that the table is correctly included as
html, this seems to fix the issue.

Another way forward would be not to include html tables here, but use
the generally accepted markdown syntax for tables. An issue is to render
the 3-column span. Maybe split tables...